+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enabled [i] environment variables as path strings
 - Changed [i] to `Invoke-Item` to the parent directory when `-Location` option is specified.
+- Added [mind2dot], [logi2dot] `-Concentrate` option
+- Enabled [json2txt] symbols for key strings
+- Changed [logi2pu] to omit `#` sign for -GrepColor option
+- Enabled [logi2dot], [logi2pu] `-Grep` to work on subgraphs
 
 ## [4.3.0]